pyproject.toml
[project]
name = "mercury-workspace"
version = "0.1.0"
description = "FactSet A2A Agent Backend and Reflex Frontend Workspace"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.135.1",
    "fds-sdk-utils>=2.1.2",
    "google-adk!=1.27.0",
    "google-cloud-aiplatform>=1.141.0",
    "google-cloud-discoveryengine>=0.13.12",
    "google-cloud-secret-manager>=2.26.0",
    "google-cloud-texttospeech>=2.34.0",
    "google-genai>=1.67.0",
    "gtts>=2.2.4",
    "mcp>=1.26.0",
    "nest-asyncio>=1.6.0",
    "plotly>=6.6.0",
    "uvicorn>=0.41.0",
]

[tool.uv]
package = false

# 🚀 NEW: Define the uv workspace members
[tool.uv.workspace]
members = ["agent", "client"]
